Squashed 'rtengine/libraw/' changes from 0b56545a4f..b860248a89
b860248a89 Changelog for 0.22.1
b59a388f5b Limit strcat space in hassy model manipulation
c0f89f30a7 Version increment; shlib increment: internal ABI has changed
69d6cfafc8 check panasonic enc8 tile width against image width
e36b26a5d9 CR3 parser: zero all buffers before fread
5cc71db156 skip memory allocation checks for OWN_ALLOC decoders; DNG SDK glue: check for memory limits; raw2image()/dcraw_process() - check for int16 source data present
fcbad80701 Check for correct bayer pattern, pass incorect ones to vng_interpolate
53365ec449 parse_rollei: zero input string before fgets
2468614a9c Nikon padded/12bit: no need to calculate padded row size before final raw_width adjustment
aa4458eb51 Fix for data size calculation integer overflow in float/deflated DNG loader (TALOS-2026-2364); Check for read results
c911c9b9ed Fix for TALOS-2026-2363: avoid integer overflow in allocation size calculation. Also: check for EOF in read loop
ac151a829b X3F decoder: implemented hard single allocation limit via LIBRAW_X3F_ALLOC_LIMIT_MB define; allocation size calculation converted to 64 bit arithm; fix for TALOS-2026-2359
b9809e410d Fix for TALOS-2026-2358
75ed2c12a3 Fix for TALOS-2026-2331
5357bb5fc6 Fix for TALOS-2026-2330
f571e44701 Sony YCC decoder: check tile size; add +3 bytes to input buffer to avoid possible overrun in huffman decoder
dae685a198 FP DNG data limit: perform calculations in 64 bit
20b7125fb0 Add extra huff_coeff item to handle huff_index==17 with known (zero) value, not externally provided tag value
0109312cb7 use %lld format for timestamp parse/print where appropriate
3e24d0740e nikon coolscan loader: check for EOF
f98a8712a5 Initialize olympus lensID bits
4484c3020e CR3 parser: all file offsets are unsigned/64bit; check current offset against file size
e628adaa5d Add Canon EOS Kiss M2 to camera list
fcb1cc0dad Check real color count against filters; do not pass really 4-color images to fbdd or advanced demosaic
95990e527a Use LIBRAW_EXCEPTION instead of own internal in losslessjpeg.h
33479aecac zero input string to avoid compare random stack garbage with tag names
cb2fa92051 Check for eof in Pentax tag search loop
24ba343e87 Fuji decoder: initialize allocated buffers

git-subtree-dir: rtengine/libraw
git-subtree-split: b860248a89d9082b8e0a1e202e516f46af9adb29

+++ src/write/file_write.cpp
@@ -195,10 +195,10 @@ void LibRaw::write_ppm_tiff()
 	{
 	    if(imgdata.params.output_flags & LIBRAW_OUTPUT_FLAGS_PPMMETA)
 	      fprintf(ofp,
-              "P7\n# EXPTIME=%0.5f\n# TIMESTAMP=%d\n# ISOSPEED=%d\n"
+              "P7\n# EXPTIME=%0.5f\n# TIMESTAMP=%lld\n# ISOSPEED=%d\n"
               "# APERTURE=%0.1f\n# FOCALLEN=%0.1f\n# MAKE=%s\n# MODEL=%s\n"
               "WIDTH %d\nHEIGHT %d\nDEPTH %d\nMAXVAL %d\nTUPLTYPE %s\nENDHDR\n",
-              shutter, (int)timestamp, (int)iso_speed,aperture, 
+              shutter, (long long int)timestamp, (int)iso_speed,aperture, 
 	      focal_len, make, model,
 	      width, height, colors, (1 << output_bps) - 1, cdesc);
 	    else
@@ -210,11 +210,11 @@ void LibRaw::write_ppm_tiff()
         else
 	{
 	    if(imgdata.params.output_flags & LIBRAW_OUTPUT_FLAGS_PPMMETA)
-	    	fprintf(ofp, "P%d\n# EXPTIME=%0.5f\n# TIMESTAMP=%d\n"
+	    	fprintf(ofp, "P%d\n# EXPTIME=%0.5f\n# TIMESTAMP=%lld\n"
 		"# ISOSPEED=%d\n# APERTURE=%0.1f\n# FOCALLEN=%0.1f\n"
 		"# MAKE=%s\n# MODEL=%s\n%d %d\n%d\n",
                 colors/2+5,
-		shutter, (int)timestamp, (int)iso_speed,aperture,focal_len,
+		shutter, (long long int)timestamp, (int)iso_speed,aperture,focal_len,
 		make,model,
 		width, height, (1 << output_bps)-1);
 	    else
